super: reduce global lock usage
Hey everyone,

This introduces a per-filesystem type spinlock to protect the list of
superblock of a filesystem instance of a given type instead of using the
global spinlock that protects the @super_blocks list.

I've had this idea for a while but now with Jan's and Christoph's rework
it's even more useful since Christoph will make it possible to drop
get_super() completely. So we're only left with few places that care
about @super_blocks.

Cc: Jens Axboe <axboe@fb.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
To: Christoph Hellwig <hch@lst.de>
To: Jan Kara <jack@suse.com>
Cc: linux-fsdevel@vger.kernel.org
Signed-off-by: Christian Brauner <brauner@kernel.org>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20230811-vfs-super-fs_supers_lock-4b6a1ccdbb72",
    "prefixes": [
      "RFC"
    ],
    "base-branch": "vfs.super.v2"
  }
}