Implement freeze and thaw as holder operations
Hey Christoph,
Hey Jan,

This implements block device freezing and thawing as holder operations.

Not just does this allow to simplify the current implementation by
removing a few locks and members in struct block_device and getting rid
of a few helpers in the superblock code, it also allows us to implement
block device freezing for multiple devices and not just the main block
device.

This will also allow us to fix block device freezing and thawing for
btrfs which is broken right now but that's just a nice side-effect.

Thanks!
Christian

To: Jan Kara <jack@suse.cz>
To: Christoph Hellwig <hch@lst.de>
Cc: linux-fsdevel@vger.kernel.org
Cc: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Christian Brauner <brauner@kernel.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20230927-vfs-super-freeze-v1-0-ecc36d9ab4d9@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20230927-vfs-super-freeze-eff650f66b06",
    "prefixes": [],
    "base-branch": "master",
    "history": {
      "v1": [
        "20230927-vfs-super-freeze-v1-0-ecc36d9ab4d9@kernel.org"
      ]
    }
  }
}