mnt_idmapping: decouple from namespaces
Hey,

This is a tiny series to fully decouple idmapped mounts from namespaces.
We already have a dedicated type and nothing matters from a namespace
apart from it's permissions. So just get rid of it. Also means we could
extend this to allow changing of idmapping completely independent of
namespaces in the future. There's no need to tie them that close
together.

Survives xfstests for btrfs, ext4, xfs and specifically the idmapped
mount tests.

Thanks!
Christian

Cc: linux-fsdevel@vger.kernel.org
Cc: Seth Forshee <sforshee@kernel.org>
Signed-off-by: Christian Brauner <brauner@kernel.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20231122-vfs-mnt_idmap-v1-0-dae4abdde5bd@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20231101-vfs-mnt_idmap-c3b7502f409d",
    "prefixes": [],
    "base-branch": "master.v6.7-rc1",
    "history": {
      "v1": [
        "20231122-vfs-mnt_idmap-v1-0-dae4abdde5bd@kernel.org"
      ]
    }
  }
}