pidfs: remove config option
Enable pidfs unconditionally. There's no real reason not do to it.
For 32bit systems we add a simple inode allocation mechanism that still
guarantees that userspace can compare processes by inode number which
they already do as I found out in [1]. If they also need the uniqueness
property that we get by default on 64bit systems they should simply
parse the contents of /proc/<pid>/fd/<nr>. On 64bit we don't have to
deal with any of this and things are nice and simple.

Link: https://github.com/systemd/systemd/pull/31713 [1]
Link: https://lore.kernel.org/r/20240312-dingo-sehnlich-b3ecc35c6de7@brauner
Signed-off-by: Christian Brauner <brauner@kernel.org>