[RFC] fs: add unmount_by_id()
* Look up the target mount with the read-side of the namespace
  semaphore held.
* Verify that the mount can either be reached from the caller's current
  root mount or that the caller is privileged in the initial namespace.
* Drop the namespace semaphore. After we dropped the namespace semaphore
  the mount layout could change again and the mount could become
  unreachable from the caller's root. That's irrelevant as similar races
  exist in the path-based umount() interface if e.g., the mount is moved
  or overmounted after it has been looked up but before we managed to
  grab the namespace semaphore.

Both statmount() and listmount() only place the requirement that the
caller must be able to to reach the target mount in order to retrieve
information about it or list it's children. In order to unmount a mount
by mount id the caller needs to be privileged over the owning user
namespace of the mount namespace of the target mount.

That restriction can be relaxed for privileged tasks. A task that has
privileges in the initial namespace should be able to unmount a mount in
any mount namespace.

Signed-off-by: Christian Brauner <brauner@kernel.org>