pidfs: allow retrieval of namespace descriptors
In recent discussions it became clear that having the ability to go from
pidfd to namespace file descriptor is desirable. Not just because it is
already possible to use pidfds with setns() to switch namespaces
atomically but also because it makes it possible to interact with
namespaces without having procfs mounted solely relying on the pidfd.

This adds support from deriving a namespace file descriptor from a
pidfd for all namespace types.

Thanks!
Christian

Cc: Josef Bacik <josef@toxicpanda.com>
Cc: Seth Forshee <sforshee@kernel.org>
Cc: Stephane Graber <stgraber@stgraber.org>
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
Cc: Aleksa Sarai <cyphar@cyphar.com>
To: linux-fsdevel@vger.kernel.org

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240627-work-pidfs-v1-0-7e9ab6cc3bb1@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240627-work-pidfs-fd415f4d3cd1",
    "prefixes": [
      "RFC"
    ],
    "base-branch": "origin-all/work.pidfs",
    "history": {
      "v1": [
        "20240627-work-pidfs-v1-0-7e9ab6cc3bb1@kernel.org"
      ]
    }
  }
}