[DRAFT]: fs: pull down audit_inode() for O_CREAT
The audit_inode() on the parent is done independent of whether the file
was actually created or not. But the audit_inode() on the file itself is
only done when it was actually created. There's no need to do
audit_inode() on the parent when the file already existed.

Audit the parent when the file was created or we failed to create it.

Signed-off-by: Christian Brauner <brauner@kernel.org>